+++ apps/meerkat/src/app/dsp/DSPConnection.ts
@@ -4,14 +4,17 @@ import {
     OperationStatistics,
     OperationInvocationInfo,
     MistypedPDUError,
+    DSABindError,
+    BindReturn,
 } from "@wildboar/meerkat-types";
 import * as errors from "@wildboar/meerkat-types";
+import { ASN1Element } from "asn1-ts";
 import { DER } from "asn1-ts/dist/node/functional";
-import type { Socket } from "net";
 import { IDMConnection } from "@wildboar/idm";
 import versions from "./versions";
 import {
     DSABindArgument,
+    _decode_DSABindArgument,
 } from "@wildboar/x500/src/lib/modules/DistributedOperations/DSABindArgument.ta";
 import {
     _encode_DSABindResult,
@@ -48,17 +51,8 @@ import {
 } from "@wildboar/x500/src/lib/modules/IDMProtocolSpecification/Abort.ta";
 import { bind as doBind } from "../authn/dsaBind";
 import {
-    dSABind,
-} from "@wildboar/x500/src/lib/modules/DistributedOperations/dSABind.oa";
-import {
     directoryBindError,
 } from "@wildboar/x500/src/lib/modules/DirectoryAbstractService/directoryBindError.oa";
-import {
-    DirectoryBindError_OPTIONALLY_PROTECTED_Parameter1,
-} from "@wildboar/x500/src/lib/modules/DirectoryAbstractService/DirectoryBindError-OPTIONALLY-PROTECTED-Parameter1.ta";
-import {
-    SecurityProblem_noInformation,
-} from "@wildboar/x500/src/lib/modules/DirectoryAbstractService/SecurityProblem.ta";
 import {
     AuthenticationLevel_basicLevels_level_none,
 } from "@wildboar/x500/src/lib/modules/BasicAccessControl/AuthenticationLevel-basicLevels-level.ta";
@@ -73,9 +67,7 @@ import * as crypto from "crypto";
 import sleep from "../utils/sleep";
 import encodeLDAPDN from "../ldap/encodeLDAPDN";
 import isDebugging from "is-debugging";
-import { dSA } from "@wildboar/x500/src/lib/modules/SelectedObjectClasses/dSA.oa";
-
-const DSA_OC_ID: string = dSA["&id"].toString();
+import { strict as assert } from "assert";
 
 async function handleRequest (
     ctx: Context,
@@ -246,14 +238,81 @@ async function handleRequestAndErrors (
 
 export default
 class DSPConnection extends ClientAssociation {
-    public readonly socket!: Socket;
+
+    public async attemptBind (arg: ASN1Element): Promise<void> {
+        const arg_ = _decode_DSABindArgument(arg);
+        const ctx = this.ctx;
+        const idm = this.idm;
+        const remoteHostIdentifier = `${idm.s.remoteFamily}://${idm.s.remoteAddress}/${idm.s.remotePort}`;
+        const startBindTime = new Date();
+        let outcome!: BindReturn;
+        try {
+            outcome = await doBind(ctx, this.idm.socket, arg_);
+        } catch (e) {
+            if (e instanceof DSABindError) {
+                ctx.log.warn(e.message);
+                const endBindTime = new Date();
+                const bindTime: number = Math.abs(differenceInMilliseconds(startBindTime, endBindTime));
+                const totalTimeInMilliseconds: number = this.ctx.config.bindMinSleepInMilliseconds
+                    + crypto.randomInt(ctx.config.bindSleepRangeInMilliseconds);
+                const sleepTime: number = Math.abs(totalTimeInMilliseconds - bindTime);
+                await sleep(sleepTime);
+                const err: typeof directoryBindError["&ParameterType"] = {
+                    unsigned: e.data,
+                };
+                const error = directoryBindError.encoderFor["&ParameterType"]!(err, BER);
+                await idm.writeBindError(dsp_ip["&id"]!, error);
+                return;
+            } else {
+                ctx.log.warn(e?.message);
+                await idm.writeAbort(Abort_reasonNotSpecified);
+                return;
+            }
+        }
+        this.boundEntry = outcome.boundVertex;
+        this.boundNameAndUID = outcome.boundNameAndUID;
+        this.authLevel = outcome.authLevel;
+        if (
+            ("basicLevels" in outcome.authLevel)
+            && (outcome.authLevel.basicLevels.level === AuthenticationLevel_basicLevels_level_none)
+        ) {
+            assert(!ctx.config.forbidAnonymousBind, "Somehow a DSA bound anonymously when anonymous binds are forbidden.");
+            ctx.log.info(ctx.i18n.t("log:connection_bound_anon", {
+                source: remoteHostIdentifier,
+                protocol: "DSP",
+            }));
+        } else {
+            ctx.log.info(ctx.i18n.t("log:connection_bound_auth", {
+                source: remoteHostIdentifier,
+                protocol: "DSP",
+                dn: this.boundNameAndUID?.dn
+                    ? encodeLDAPDN(ctx, this.boundNameAndUID.dn)
+                    : "",
+            }));
+        }
+        const bindResult = new DSABindArgument(
+            undefined, // TODO: Supply return credentials. NOTE that the specification says that this must be the same CHOICE that the user supplied.
+            versions,
+        );
+        await idm.writeBindResult(dsp_ip["&id"]!, _encode_DSABindResult(bindResult, BER));
+        idm.events.removeAllListeners("request");
+        idm.events.on("request", this.handleRequest.bind(this));
+    }
 
     private async handleRequest (request: Request): Promise<void> {
         await handleRequestAndErrors(this.ctx, this, request);
     }
 
     private async handleUnbind (): Promise<void> {
-        await this.ctx.db.enqueuedSearchResult.deleteMany({
+        if (!this.isBound()) {
+            return; // We don't want users to be able to spam unbinds.
+        }
+        this.ctx.db.enqueuedListResult.deleteMany({ // INTENTIONAL_NO_AWAIT
+            where: {
+                connection_uuid: this.id,
+            },
+        });
+        this.ctx.db.enqueuedSearchResult.deleteMany({ // INTENTIONAL_NO_AWAIT
             where: {
                 connection_uuid: this.id,
             },
@@ -268,111 +327,10 @@ class DSPConnection extends ClientAssociation {
     constructor (
         readonly ctx: Context,
         readonly idm: IDMConnection,
-        readonly bind: DSABindArgument,
     ) {
         super();
-        if (!ctx.config.dsp.enabled) {
-            idm.writeAbort(Abort_invalidProtocol).then(() => idm.events.emit("unbind", null));
-            return;
-        }
         this.socket = idm.s;
-        const remoteHostIdentifier = `${idm.s.remoteFamily}://${idm.s.remoteAddress}/${idm.s.remotePort}`;
-        const startBindTime = new Date();
-        doBind(ctx, idm.s, bind)
-            .then(async (outcome) => {
-                const inappropriateBoundEntry: boolean = (
-                    !ctx.config.dsaCanBindAsNonDSA
-                    && !outcome.boundVertex?.dse.objectClass.has(DSA_OC_ID)
-                );
-                if (outcome.failedAuthentication || inappropriateBoundEntry) {
-                    const endBindTime = new Date();
-                    const bindTime: number = Math.abs(differenceInMilliseconds(startBindTime, endBindTime));
-                    const totalTimeInMilliseconds: number = ctx.config.bindMinSleepInMilliseconds
-                        + crypto.randomInt(ctx.config.bindSleepRangeInMilliseconds);
-                    const sleepTime: number = Math.abs(totalTimeInMilliseconds - bindTime);
-                    await sleep(sleepTime);
-                }
-                this.boundEntry = outcome.boundVertex;
-                this.boundNameAndUID = outcome.boundNameAndUID;
-                this.authLevel = outcome.authLevel;
-                if (outcome.failedAuthentication || inappropriateBoundEntry) {
-                    const err: typeof directoryBindError["&ParameterType"] = {
-                        unsigned: new DirectoryBindError_OPTIONALLY_PROTECTED_Parameter1(
-                            versions,
-                            {
-                                securityError: SecurityProblem_noInformation,
-                            },
-                            undefined, // Failed authentication will not yield any security parameters.
-                        ),
-                    };
-                    const error = dSABind.encoderFor["&ParameterType"]!(err, BER);
-                    idm
-                        .writeBindError(dsp_ip["&id"]!, error)
-                        .then(() => {
-                            this.handleUnbind()
-                                .then(() => {
-                                    ctx.log.info(ctx.i18n.t("log:disconnected_auth_failure", {
-                                        ctype: DSPConnection.name,
-                                        cid: this.id,
-                                        source: remoteHostIdentifier,
-                                    }));
-                                })
-                                .catch((e) => {
-                                    ctx.log.info(ctx.i18n.t("log:error_unbinding", {
-                                        ctype: DSPConnection.name,
-                                        cid: this.id,
-                                        e: e.message,
-                                    }));
-                                });
-                        })
-                        .catch((e) => {
-                            ctx.log.info(ctx.i18n.t("log:error_writing_bind_error", {
-                                ctype: DSPConnection.name,
-                                cid: this.id,
-                                e: e.message,
-                            }));
-                        });
-                    ctx.log.info(ctx.i18n.t("log:auth_failure", {
-                        ctype: DSPConnection.name,
-                        cid: this.id,
-                        source: remoteHostIdentifier,
-                    }));
-                    return;
-                }
-                const bindResult = new DSABindArgument( // DSABindResult === DSABindArgument
-                    undefined,
-                    versions,
-                );
-                idm.writeBindResult(dsp_ip["&id"]!, _encode_DSABindResult(bindResult, BER));
-                if (
-                    ("basicLevels" in outcome.authLevel)
-                    && (outcome.authLevel.basicLevels.level === AuthenticationLevel_basicLevels_level_none)
-                ) {
-                    ctx.log.info(ctx.i18n.t("log:connection_bound_anon", {
-                        source: remoteHostIdentifier,
-                        protocol: "DSP",
-                    }));
-                } else {
-                    ctx.log.info(ctx.i18n.t("log:connection_bound_auth", {
-                        source: remoteHostIdentifier,
-                        protocol: "DSP",
-                        dn: this.boundNameAndUID?.dn
-                            ? encodeLDAPDN(ctx, this.boundNameAndUID.dn)
-                            : "",
-                    }));
-                }
-            })
-            .catch((e) => {
-                ctx.log.error(ctx.i18n.t("log:bind_error", {
-                    ctype: DSPConnection.name,
-                    cid: this.id,
-                    source: remoteHostIdentifier,
-                    e: e.message,
-                }));
-            });
-
-        idm.events.removeAllListeners("request");
-        idm.events.on("request", this.handleRequest.bind(this));
+        assert(ctx.config.dsp.enabled, "User somehow bound via DSP when it was disabled.");
         idm.events.on("unbind", this.handleUnbind.bind(this));
     }
 }
